// src/pages/previous-games/previous-games.enum.ts
export enum PreviousGamesEnum {

    PLATFORM_IOS = 'ios',
    PLATFORM_ANDROID = 'android',
    PLATFORM_WINDOWS = 'windows',

    SHARED_PREFERENCE = 'pre_defined',

    OFFLINE_MESSAGE = 'Your internet connection appears to be offline, please try again later.',
    ONLINE_MESSAGE = 'You are back online',

    LINK = 'assets/imgs/golf-home.jpg',

    TOAST_OKAY = 'Ok',
    TOAST_CSS = "toast-container",

    ALERT_OKAY = 'Ok',
    ALERT_TITLE = 'Selecting Date Range',

    LOADER_CONTENT = 'Loading Data',

    PREVIOUS_GAME_DETAIL_PAGE = 'PreviousGameDetailPage'
}